+++ README.md
@@ -113,8 +113,9 @@ Currently, the library contains the following features:
 * ARIMA & auto-ARIMA,
 * Facebook Prophet,
 * Theta method,
-* FFT (Fast Fourier Transform)
-* Recurrent neural networks (vanilla RNNs, GRU, and LSTM variants).
+* FFT (Fast Fourier Transform),
+* Recurrent neural networks (vanilla RNNs, GRU, and LSTM variants),
+* Temporal convolutional network.
 
 **Preprocessing:** Transformer tool for easily scaling / normalizing time series.
