+++ u8timeseries/models/__init__.py
@@ -9,6 +9,7 @@ from .baselines import NaiveMean, NaiveSeasonal, NaiveDrift
 from .prophet import Prophet
 from .exponential_smoothing import ExponentialSmoothing
 from .rnn_model import RNNModel
+from .tcn_model import TCNModel
 from .theta import Theta
 from .fft import FFT
