+++ u8timeseries/tests/test_dataset.py
@@ -3,7 +3,7 @@ import pandas as pd
 import torch
 
 from ..timeseries import TimeSeries
-from ..models.rnn_model import _TimeSeriesDataset1D
+from ..models.torch_forecasting_model import _TimeSeriesDataset1DSequential
 
 
 class DatasetTestCase(unittest.TestCase):
@@ -15,14 +15,14 @@ class DatasetTestCase(unittest.TestCase):
     def test_creation(self):
         # Cannot have train window <= 0
         with self.assertRaises(ValueError):
-            _TimeSeriesDataset1D(self.series, -1)
+            _TimeSeriesDataset1DSequential(self.series, -1)
         # Cannot have label window <= 0
         with self.assertRaises(ValueError):
-            _TimeSeriesDataset1D(self.series, 1, -1)
+            _TimeSeriesDataset1DSequential(self.series, 1, -1)
 
     def test_content(self):
         # Can have no transformation
-        dataset = _TimeSeriesDataset1D(self.series, 12, 4)
+        dataset = _TimeSeriesDataset1DSequential(self.series, 12, 4)
         self.assertEqual(len(dataset), 100 - 12 - 4 + 1)
         # correct types
         self.assertEqual(type(dataset[0][0]), torch.Tensor)
